+++ ear-production-suite-plugins/plugins/direct_speakers/CMakeLists.txt
@@ -49,7 +49,7 @@ set(HEADERS_DIRECT_SPEAKERS_INPUT
     src/direct_speakers_plugin_editor.hpp
     src/direct_speakers_plugin_processor.hpp
     src/direct_speakers_component.hpp
-    src/channel_gains_box.hpp
+		src/speaker_meters_box.hpp
 		src/speaker_level.hpp
     src/value_box_main.hpp
     src/value_box_metadata.hpp
